requirements.txt
requests
cryptography
azure-storage-blob